requirements.txt

viam-sdk==0.43.0

typing-extensions
psutil